@@ -27,10 +27,12 @@
 		"tauri": "tauri"
 	},
 	"devDependencies": {
-		"@babel/plugin-proposal-decorators": "^7.25.7",
-		"@sveltejs/adapter-auto": "^3.2.3",
-		"@sveltejs/adapter-static": "^3.0.3",
-		"@sveltejs/kit": "^2.5.21",
+		"svelte": "^5.0.3",
+		"@sveltejs/adapter-auto": "^3.2.5",
+		"@sveltejs/adapter-static": "^3.0.%",
+		"@sveltejs/kit": "^2.7.2",
+		"eslint-plugin-svelte": "^2.45.1",
+		"prettier-plugin-svelte": "^3.2.7",
 		"@tauri-apps/cli": "^2.0.1",
 		"@typescript-eslint/eslint-plugin": "^8.9.0",
 		"@typescript-eslint/parser": "^8.9.0",
@@ -38,12 +40,9 @@
 		"dotenv": "^16.4.5",
 		"eslint": "^9.12.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.45.1",
 		"install": "^0.13.0",
 		"prettier": "^3.3.3",
-		"prettier-plugin-svelte": "^3.2.7",
 		"sass": "^1.77.8",
-		"svelte": "^4.2.18",
 		"typescript": "^5.6.3",
 		"typescript-eslint": "^8.1.0",
 		"vite": "^5.4.9",
