@@ -1,4 +1,4 @@
-import { writable } from 'svelte/store';
+import { get, writable } from 'svelte/store';
 import { browser } from '$app/environment';
 import { Preferences } from '@capacitor/preferences';
 
@@ -7,11 +7,13 @@ export const createStore = (key: string, defaultValue: string) => {
 
 	if (browser) {
 		Preferences.get({ key }).then(({ value }) => {
+			// console.log('got value from preferences', value);
 			capacitorValue = value;
 		});
 	}
 
-	const store = writable<string>(browser ? capacitorValue ?? defaultValue : defaultValue);
+	const store = writable<string>(browser ? (capacitorValue ?? defaultValue) : defaultValue);
+	// console.log(get(store));
 
 	store.subscribe((value) => {
 		if (browser) {
