@@ -22,9 +22,10 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 	import { Browser } from '@capacitor/browser';
 	import { Capacitor } from '@capacitor/core';
-	import { Preferences } from '@capacitor/preferences';
 	import { StatusBar, Style } from '@capacitor/status-bar';
 	import toast from 'svelte-french-toast';
+
+	import { Preferences } from '$lib/plugins/Preferences';
 
 	import { dropboxLogout, getDropboxAuthUrl } from '$lib/driver/PoemDropboxStorageDriver';
 	import { activeLanguage } from '$lib/stores/activeLanguage';
